handle url- and downloadurl-field according to SSZ
resource_dict never has a download_url field, this only exists in DCAT. To know if it's a download or not, you have to check the format. xml, wms, wmts, wfs, rest are services, and thus don't need a downloadURL field, everything else is a download, where downloadURL and accessURL is the same.