Create resource views in before_view
IDataPusher after_upload would actually be the correct place to
implement this functionality. But for a yet unknown reason on the INTEG
environment DataPusher fails to call the result_url (yet the process
finished successfully) and therefore after_upload never gets called.
Therefore this hack needs to be implemented and the resource views must
be created when a dataset is being prepared to be displayed.

Hopefully this can be removed in the future and the proper solution
using the IDataPusher interface can be used again.