Require six
ckanext-showcase requires six>=1.12.0 in its setup.py. This causes a failure when installing it for the Travis run, likely because of the Let's Encrypt certificate issue meaning that the Travis runner cannot access pypi to find a fitting version to download. (Installing directly with pip does not run into this issue.) Therefore, we require a fitting version of six in our ckanext and make sure it is installed before installing ckanext-showcase.