feat: Fix the tiling of the group items
In CKAN 2.9, the maximum width of the main container has been increased:
@media (min-width: 992px) {
  .container {
    width: 970px;
  }
}
@media (min-width: 1200px) {
  .container {
    width: 1170px;
  }
}

Previously the maximum width of the container was 940px. So that the group item tiles still look all right and we show six in a row on a desktop screen, I have adjusted the css.

I also got fed up with the simple masonry implementation in the ckan data-module 'media-grid', and decided to ignore it and use flexbox instead.

The masonry.pkgd.js asset that was here before was added temporarily because the masonry library in CKAN 2.7 had a bug. It can now be removed completely.