fix: Handle cases where a dataset field doesn't exist
Specifically, this fixes the case where some old datasets didn't have the field 'updateInterval', which should contain a dict of strings. When we got back the default value of None and tried to find the first element of that, we got a TypeError.

Where we can't handle the default return value of None from _get_dataset_value, we need to pass in a default value that we can handle.