chore: Remove unneeded try/except
The code inside the try block has changed so that we can no longer get an IndexError from it, so no need to catch one. (Tags in the tag list should always have the key 'name', but I changed this to tag.get('name') instead of tag['name'] just in case.