fix: Remove unneeded configure method
This method was being called when running 'ckan db init' and causing an error, because it looked for the site user at a time when the db had not yet been brought up and had no users table.

We should not need to create the tag vocabularies at this stage - it should be enough to have the methods for creating them used as converters for their fields in the package schema (which they are).